requirements.txt
pip
SQLAlchemy
Flask-SQLAlchemy
mysql-python
tornado
Flask-RESTful
numpy
scikit-learn
scipy
requests
